Move into hlrc and fix comment
